refactor(edit-engine): extract Instance transaction domains
Move Instance lifecycle and Symbol changes, placement transforms, and netlist identity/parameter edits into three typed dispatchers. Each dispatcher receives only the draft, resolver or route-follow state it needs, change tracking, and the per-edit rejection seam.

Preserve ordered draft mutation, exact rejection paths/object IDs, deferred Net pruning, annotation and Route following, layout-lock policy, connectivity flags, and reference validation. executeTransaction remains the sole public mutation entry and still owns post-transform direct-contact reconciliation.

Validation: pnpm typecheck; pnpm test:local packages/edit-engine/src plus Gallery placement, selection deletion, and clipboard contracts (271 tests); git diff --check.

No test files changed in this extraction because the permanent transaction, hierarchy, Gallery placement, deletion, clipboard, and transform contracts already exercise the moved behavior.

Test-Impact: no-test-change — behavior-preserving typed dispatcher extraction covered by 271 focused tests.
