+++ packages/edit-engine/src/transaction-instance-netlist.ts
@@ -1,0 +1,563 @@
+import { createReferenceIndex, referenceIssuesForInstance } from "@icm/devices";
+import type { SchematicDocument } from "@icm/model";
+
+import type { EditTransaction } from "./edit-schema.js";
+import type { RejectEdit } from "./transaction-cell-interface.js";
+import { refreshInstanceValueAnnotation } from "./transaction-instance-annotations.js";
+import type { RejectedTransaction } from "./transaction-result.js";
+
+type InstanceNetlistEdit = Extract<
+  EditTransaction["edits"][number],
+  {
+    kind:
+      | "patch_instance_netlist_parameters"
+      | "set_instance_reference"
+      | "set_instance_schematic_reference"
+      | "set_instance_schematic_name"
+      | "set_instance_binding"
+      | "set_instance_netlist"
+      | "bulk_patch_instance_netlist";
+  }
+>;
+
+export interface InstanceNetlistEditContext {
+  draft: SchematicDocument;
+  changedObjectIds: Set<string>;
+  reject: RejectEdit;
+}
+
+export type InstanceNetlistEditOutcome =
+  | { ok: true; connectivityChanged: boolean }
+  | { ok: false; rejection: RejectedTransaction };
+
+function referencePolicyFailure(
+  draft: SchematicDocument,
+  instanceId: string,
+): string | null {
+  const issue = referenceIssuesForInstance(
+    createReferenceIndex(draft),
+    instanceId,
+  )[0];
+  if (!issue) return null;
+  switch (issue.code) {
+    case "MISSING_REFERENCE":
+      return "This component requires a netlist reference";
+    case "UNEXPECTED_REFERENCE":
+      return "This symbol does not emit a netlist reference";
+    case "WRONG_REFERENCE_PREFIX":
+      return `Reference ${issue.reference} does not match this component prefix`;
+    case "DUPLICATE_REFERENCE":
+      return `Reference ${issue.reference} is already used by ${issue.otherInstanceId}`;
+  }
+}
+
+export function applyInstanceNetlistEdit(
+  edit: InstanceNetlistEdit,
+  context: InstanceNetlistEditContext,
+): InstanceNetlistEditOutcome {
+  const { draft, changedObjectIds, reject } = context;
+
+  switch (edit.kind) {
+    case "patch_instance_netlist_parameters": {
+      const instance = draft.instances.find(
+        (candidate) => candidate.id === edit.instanceId,
+      );
+      if (!instance) {
+        return {
+          ok: false,
+          rejection: reject(
+            "OBJECT_NOT_FOUND",
+            `Instance does not exist: ${edit.instanceId}`,
+            [],
+            [edit.instanceId],
+          ),
+        };
+      }
+      const set = edit.set ?? {};
+      const unset = edit.unset ?? [];
+      if (Object.keys(set).length === 0 && unset.length === 0) {
+        return {
+          ok: false,
+          rejection: reject(
+            "EDIT_PRECONDITION",
+            "Netlist parameter patch must set or unset at least one parameter",
+            [],
+            [edit.instanceId],
+          ),
+        };
+      }
+      const duplicateUnset = new Set(unset);
+      if (duplicateUnset.size !== unset.length) {
+        return {
+          ok: false,
+          rejection: reject(
+            "EDIT_PRECONDITION",
+            "Netlist parameter patch cannot unset the same parameter more than once",
+            [],
+            [edit.instanceId],
+          ),
+        };
+      }
+      const conflictingKey = Object.keys(set).find((key) =>
+        duplicateUnset.has(key),
+      );
+      if (conflictingKey) {
+        return {
+          ok: false,
+          rejection: reject(
+            "EDIT_PRECONDITION",
+            `Netlist parameter patch cannot set and unset ${conflictingKey}`,
+            [],
+            [edit.instanceId],
+          ),
+        };
+      }
+      const before: SchematicDocument["instances"][number] =
+        structuredClone(instance);
+      const netlist = instance.netlist;
+      if (!netlist) {
+        return {
+          ok: false,
+          rejection: reject(
+            "EDIT_PRECONDITION",
+            "Netlist parameter patch requires an instance netlist record",
+            [],
+            [edit.instanceId],
+          ),
+        };
+      }
+      const nextParameters = { ...netlist.parameters };
+      for (const key of unset) delete nextParameters[key];
+      for (const [key, value] of Object.entries(set)) {
+        nextParameters[key] = value;
+      }
+      const namesByFoldedName = new Map<string, string>();
+      for (const name of Object.keys(nextParameters)) {
+        const foldedName = name.toLowerCase();
+        const prior = namesByFoldedName.get(foldedName);
+        if (prior && prior !== name) {
+          return {
+            ok: false,
+            rejection: reject(
+              "EDIT_PRECONDITION",
+              `Netlist parameter ${name} duplicates ${prior} under case folding`,
+              [],
+              [edit.instanceId],
+            ),
+          };
+        }
+        namesByFoldedName.set(foldedName, name);
+      }
+      const changed =
+        Object.keys(nextParameters).length !==
+          Object.keys(netlist.parameters).length ||
+        Object.entries(nextParameters).some(
+          ([key, value]) => netlist.parameters[key] !== value,
+        );
+      if (!changed) {
+        return {
+          ok: false,
+          rejection: reject(
+            "EDIT_PRECONDITION",
+            "Netlist parameter patch does not change the instance",
+            [],
+            [edit.instanceId],
+          ),
+        };
+      }
+      netlist.parameters = nextParameters;
+      refreshInstanceValueAnnotation(
+        draft,
+        before,
+        edit.instanceId,
+        changedObjectIds,
+      );
+      changedObjectIds.add(edit.instanceId);
+      return { ok: true, connectivityChanged: false };
+    }
+    case "set_instance_reference": {
+      const instance = draft.instances.find(
+        (candidate) => candidate.id === edit.instanceId,
+      );
+      if (!instance?.netlist) {
+        return {
+          ok: false,
+          rejection: reject(
+            "EDIT_PRECONDITION",
+            "Reference edit requires an instance netlist record",
+            [],
+            [edit.instanceId],
+          ),
+        };
+      }
+      if (instance.netlist.reference === edit.reference) {
+        return {
+          ok: false,
+          rejection: reject(
+            "EDIT_PRECONDITION",
+            "Reference edit does not change the instance",
+            [],
+            [edit.instanceId],
+          ),
+        };
+      }
+      instance.netlist.reference = edit.reference;
+      const failure = referencePolicyFailure(draft, instance.id);
+      if (failure) {
+        return {
+          ok: false,
+          rejection: reject("EDIT_PRECONDITION", failure, [], [instance.id]),
+        };
+      }
+      changedObjectIds.add(edit.instanceId);
+      return { ok: true, connectivityChanged: false };
+    }
+    case "set_instance_schematic_reference": {
+      const instance = draft.instances.find(
+        (candidate) => candidate.id === edit.instanceId,
+      );
+      if (!instance) {
+        return {
+          ok: false,
+          rejection: reject(
+            "OBJECT_NOT_FOUND",
+            `Instance does not exist: ${edit.instanceId}`,
+            [],
+            [edit.instanceId],
+          ),
+        };
+      }
+      if (
+        draft.netlist?.terminals.some((terminal) =>
+          terminal.interfaceInstanceIds.includes(instance.id),
+        )
+      ) {
+        return {
+          ok: false,
+          rejection: reject(
+            "EDIT_PRECONDITION",
+            "A formal Cell Pin is identified by its Cell terminal name, not a schematic reference",
+            [],
+            [instance.id],
+          ),
+        };
+      }
+      if (instance.schematicReference === edit.reference) {
+        return {
+          ok: false,
+          rejection: reject(
+            "EDIT_PRECONDITION",
+            "Schematic reference edit does not change the instance",
+            [],
+            [edit.instanceId],
+          ),
+        };
+      }
+      const duplicate = draft.instances.find(
+        (candidate) =>
+          candidate.id !== instance.id &&
+          (
+            candidate.schematicReference ?? candidate.netlist?.reference
+          )?.toLowerCase() === edit.reference.toLowerCase(),
+      );
+      if (duplicate) {
+        return {
+          ok: false,
+          rejection: reject(
+            "EDIT_PRECONDITION",
+            `Schematic reference is already in use: ${edit.reference}`,
+            [],
+            [duplicate.id],
+          ),
+        };
+      }
+      instance.schematicReference = edit.reference;
+      changedObjectIds.add(instance.id);
+      for (const annotation of draft.annotations) {
+        if (
+          annotation.binding?.kind === "instance-schematic-name" &&
+          annotation.binding.instanceId === instance.id
+        ) {
+          changedObjectIds.add(annotation.id);
+        }
+      }
+      return { ok: true, connectivityChanged: false };
+    }
+    case "set_instance_schematic_name": {
+      const instance = draft.instances.find(
+        (candidate) => candidate.id === edit.instanceId,
+      );
+      if (!instance) {
+        return {
+          ok: false,
+          rejection: reject(
+            "OBJECT_NOT_FOUND",
+            `Instance does not exist: ${edit.instanceId}`,
+            [],
+            [edit.instanceId],
+          ),
+        };
+      }
+      if (
+        JSON.stringify(instance.schematicName ?? null) ===
+        JSON.stringify(edit.content)
+      ) {
+        return {
+          ok: false,
+          rejection: reject(
+            "EDIT_PRECONDITION",
+            "Schematic name edit does not change the instance",
+            [],
+            [edit.instanceId],
+          ),
+        };
+      }
+      instance.schematicName = structuredClone(edit.content);
+      changedObjectIds.add(edit.instanceId);
+      for (const annotation of draft.annotations) {
+        if (
+          annotation.binding?.kind === "instance-schematic-name" &&
+          annotation.binding.instanceId === instance.id
+        ) {
+          changedObjectIds.add(annotation.id);
+        }
+      }
+      return { ok: true, connectivityChanged: false };
+    }
+    case "set_instance_binding": {
+      const instance = draft.instances.find(
+        (candidate) => candidate.id === edit.instanceId,
+      );
+      if (!instance?.netlist) {
+        return {
+          ok: false,
+          rejection: reject(
+            "EDIT_PRECONDITION",
+            "Binding edit requires an instance netlist record",
+            [],
+            [edit.instanceId],
+          ),
+        };
+      }
+      const current = instance.netlist.binding ?? null;
+      if (JSON.stringify(current) === JSON.stringify(edit.binding)) {
+        return {
+          ok: false,
+          rejection: reject(
+            "EDIT_PRECONDITION",
+            "Binding edit does not change the instance",
+            [],
+            [edit.instanceId],
+          ),
+        };
+      }
+      if (edit.binding)
+        instance.netlist.binding = structuredClone(edit.binding);
+      else delete instance.netlist.binding;
+      const failure = referencePolicyFailure(draft, instance.id);
+      if (failure) {
+        return {
+          ok: false,
+          rejection: reject("EDIT_PRECONDITION", failure, [], [instance.id]),
+        };
+      }
+      changedObjectIds.add(edit.instanceId);
+      return { ok: true, connectivityChanged: true };
+    }
+    case "set_instance_netlist": {
+      const instance = draft.instances.find(
+        (candidate) => candidate.id === edit.instanceId,
+      );
+      if (!instance) {
+        return {
+          ok: false,
+          rejection: reject(
+            "OBJECT_NOT_FOUND",
+            `Instance does not exist: ${edit.instanceId}`,
+            [],
+            [edit.instanceId],
+          ),
+        };
+      }
+      if (
+        instance.netlist &&
+        JSON.stringify(instance.netlist) === JSON.stringify(edit.netlist)
+      ) {
+        return {
+          ok: false,
+          rejection: reject(
+            "EDIT_PRECONDITION",
+            "Netlist edit does not change the instance",
+            [],
+            [edit.instanceId],
+          ),
+        };
+      }
+      const before: SchematicDocument["instances"][number] =
+        structuredClone(instance);
+      instance.netlist = structuredClone(edit.netlist);
+      const failure = referencePolicyFailure(draft, instance.id);
+      if (failure) {
+        return {
+          ok: false,
+          rejection: reject("EDIT_PRECONDITION", failure, [], [instance.id]),
+        };
+      }
+      refreshInstanceValueAnnotation(
+        draft,
+        before,
+        edit.instanceId,
+        changedObjectIds,
+      );
+      changedObjectIds.add(edit.instanceId);
+      return { ok: true, connectivityChanged: true };
+    }
+    case "bulk_patch_instance_netlist": {
+      const assignedIds = new Set<string>();
+      let connectivityChanged = false;
+      for (const assignment of edit.assignments) {
+        if (assignedIds.has(assignment.instanceId)) {
+          return {
+            ok: false,
+            rejection: reject(
+              "EDIT_PRECONDITION",
+              `Bulk netlist patch repeats instance ${assignment.instanceId}`,
+              [],
+              [assignment.instanceId],
+            ),
+          };
+        }
+        assignedIds.add(assignment.instanceId);
+        const instance = draft.instances.find(
+          (candidate) => candidate.id === assignment.instanceId,
+        );
+        if (!instance?.netlist) {
+          return {
+            ok: false,
+            rejection: reject(
+              "EDIT_PRECONDITION",
+              `Bulk netlist patch requires a netlist record: ${assignment.instanceId}`,
+              [],
+              [assignment.instanceId],
+            ),
+          };
+        }
+        const before: SchematicDocument["instances"][number] =
+          structuredClone(instance);
+        let changed = false;
+        let parametersChanged = false;
+        if (
+          assignment.reference !== undefined &&
+          instance.netlist.reference !== assignment.reference
+        ) {
+          instance.netlist.reference = assignment.reference;
+          changed = true;
+        }
+        if (assignment.binding !== undefined) {
+          const current = instance.netlist.binding ?? null;
+          if (JSON.stringify(current) !== JSON.stringify(assignment.binding)) {
+            if (assignment.binding) {
+              instance.netlist.binding = structuredClone(assignment.binding);
+            } else {
+              delete instance.netlist.binding;
+            }
+            changed = true;
+            connectivityChanged = true;
+          }
+        }
+        const set = assignment.set ?? {};
+        const unset = assignment.unset ?? [];
+        const unsetNames = new Set(unset.map((name) => name.toLowerCase()));
+        if (unsetNames.size !== unset.length) {
+          return {
+            ok: false,
+            rejection: reject(
+              "EDIT_PRECONDITION",
+              `Bulk netlist patch repeats an unset parameter on ${instance.id}`,
+              [],
+              [instance.id],
+            ),
+          };
+        }
+        const conflictingKey = Object.keys(set).find((key) =>
+          unsetNames.has(key.toLowerCase()),
+        );
+        if (conflictingKey) {
+          return {
+            ok: false,
+            rejection: reject(
+              "EDIT_PRECONDITION",
+              `Bulk netlist patch cannot set and unset ${conflictingKey}`,
+              [],
+              [instance.id],
+            ),
+          };
+        }
+        if (Object.keys(set).length > 0 || unset.length > 0) {
+          const nextParameters = { ...instance.netlist.parameters };
+          for (const key of unset) delete nextParameters[key];
+          for (const [key, value] of Object.entries(set)) {
+            nextParameters[key] = value;
+          }
+          const namesByFoldedName = new Map<string, string>();
+          for (const name of Object.keys(nextParameters)) {
+            const folded = name.toLowerCase();
+            const prior = namesByFoldedName.get(folded);
+            if (prior && prior !== name) {
+              return {
+                ok: false,
+                rejection: reject(
+                  "EDIT_PRECONDITION",
+                  `Netlist parameter ${name} duplicates ${prior} under case folding`,
+                  [],
+                  [instance.id],
+                ),
+              };
+            }
+            namesByFoldedName.set(folded, name);
+          }
+          parametersChanged =
+            Object.keys(nextParameters).length !==
+              Object.keys(instance.netlist.parameters).length ||
+            Object.entries(nextParameters).some(
+              ([key, value]) => instance.netlist!.parameters[key] !== value,
+            );
+          if (parametersChanged) {
+            instance.netlist.parameters = nextParameters;
+            changed = true;
+          }
+        }
+        if (!changed) {
+          return {
+            ok: false,
+            rejection: reject(
+              "EDIT_PRECONDITION",
+              `Bulk netlist patch does not change ${instance.id}`,
+              [],
+              [instance.id],
+            ),
+          };
+        }
+        if (parametersChanged) {
+          refreshInstanceValueAnnotation(
+            draft,
+            before,
+            instance.id,
+            changedObjectIds,
+          );
+        }
+        changedObjectIds.add(instance.id);
+      }
+      for (const instanceId of assignedIds) {
+        const failure = referencePolicyFailure(draft, instanceId);
+        if (failure) {
+          return {
+            ok: false,
+            rejection: reject("EDIT_PRECONDITION", failure, [], [instanceId]),
+          };
+        }
+      }
+      return { ok: true, connectivityChanged };
+    }
+  }
+}
