+++ packages/edit-engine/src/transaction.ts
@@ -1,13 +1,10 @@
 import {
   AnnotationSchema,
   ConnectivityEvidenceSchema,
-  InstanceSchema,
   JunctionSchema,
-  NoConnectSchema,
   SchematicDocumentSchema,
   deriveStableId,
 } from "@icm/model";
-import { createReferenceIndex, referenceIssuesForInstance } from "@icm/devices";
 import type { RouteBranch, SchematicDocument } from "@icm/model";
 import {
   endpointKey,
@@ -40,18 +37,14 @@ import {
   type NetLabelRouteAnchor,
   type RouteMarkerAnchor,
 } from "./transaction-route-annotations.js";
-import {
-  applyInstanceRouteFollow,
-  splitRoute,
-} from "./transaction-route-follow.js";
-import {
-  followAttachedAnnotations,
-  refreshInstanceValueAnnotation,
-} from "./transaction-instance-annotations.js";
+import { splitRoute } from "./transaction-route-follow.js";
 import { reconcileTransformDirectContacts } from "./transaction-direct-contact.js";
 import { nextPhysicalContactOperation } from "./transaction-connectivity-normalizer.js";
 import { applyCellResetEdit } from "./transaction-cell-reset.js";
 import { applyCellInterfaceEdit } from "./transaction-cell-interface.js";
+import { applyInstanceLifecycleEdit } from "./transaction-instance-lifecycle.js";
+import { applyInstanceNetlistEdit } from "./transaction-instance-netlist.js";
+import { applyInstanceTransformEdit } from "./transaction-instance-transform.js";
 import { applyPresentationLayoutEdit } from "./transaction-presentation-layout.js";
 import {
   type BulkDefaultIdentity,
@@ -100,27 +93,6 @@ import {
 
 export * from "./edit-schema.js";
 export * from "./transaction-result.js";
-
-function referencePolicyFailure(
-  draft: SchematicDocument,
-  instanceId: string,
-): string | null {
-  const issue = referenceIssuesForInstance(
-    createReferenceIndex(draft),
-    instanceId,
-  )[0];
-  if (!issue) return null;
-  switch (issue.code) {
-    case "MISSING_REFERENCE":
-      return "This component requires a netlist reference";
-    case "UNEXPECTED_REFERENCE":
-      return "This symbol does not emit a netlist reference";
-    case "WRONG_REFERENCE_PREFIX":
-      return `Reference ${issue.reference} does not match this component prefix`;
-    case "DUPLICATE_REFERENCE":
-      return `Reference ${issue.reference} is already used by ${issue.otherInstanceId}`;
-  }
-}
 
 export function executeTransaction(
   document: SchematicDocument,
@@ -272,948 +244,51 @@ export function executeTransaction(
         geometryChanged ||= outcome.geometryChanged;
         break;
       }
-      case "add_instance": {
-        const objectIdExists = [
-          ...draft.instances,
-          ...draft.nets,
-          ...draft.routes,
-          ...draft.junctions,
-          ...draft.noConnects,
-          ...draft.annotations,
-          ...draft.layoutGroups,
-          ...draft.constraints,
-        ].some((candidate) => candidate.id === edit.instance.id);
-        if (objectIdExists) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            `Object ID already exists: ${edit.instance.id}`,
-          );
-        }
-        const resolver = context.symbolResolver;
-        if (
-          !resolver?.resolve(
-            edit.instance.symbolId,
-            edit.instance.symbolVariantId,
-          )
-        ) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            `Symbol does not exist: ${edit.instance.symbolId}`,
-          );
-        }
-        draft.instances.push(InstanceSchema.parse(edit.instance));
-        changedObjectIds.add(edit.instance.id);
-        connectivityChanged = true;
-        break;
-      }
-      case "remove_instance": {
-        const index = draft.instances.findIndex(
-          (candidate) => candidate.id === edit.instanceId,
-        );
-        if (index < 0) {
-          return rejectAt(
-            "OBJECT_NOT_FOUND",
-            `Instance does not exist: ${edit.instanceId}`,
-            [],
-            [edit.instanceId],
-          );
-        }
-        const referenced =
-          draft.nets.some((net) =>
-            net.terminals.some(
-              (terminal) => terminal.instanceId === edit.instanceId,
-            ),
-          ) ||
-          draft.noConnects.some(
-            (noConnect) =>
-              noConnect.endpoint.kind === "terminal" &&
-              noConnect.endpoint.instanceId === edit.instanceId,
-          ) ||
-          draft.annotations.some(
-            (annotation) =>
-              annotation.anchor.kind === "object" &&
-              annotation.anchor.objectId === edit.instanceId,
-          ) ||
-          draft.layoutGroups.some((group) =>
-            group.objectIds.includes(edit.instanceId),
-          ) ||
-          draft.constraints.some((constraint) =>
-            constraint.objectIds.includes(edit.instanceId),
-          );
-        if (referenced) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            `Instance is still connected or referenced: ${edit.instanceId}`,
-          );
-        }
-        const ownerNetIds = removeConnectivityEvidenceOwnedBy(
-          draft,
-          new Set([edit.instanceId]),
-          changedObjectIds,
-        );
-        draft.instances.splice(index, 1);
-        changedObjectIds.add(edit.instanceId);
-        for (const netId of ownerNetIds) {
-          deferNetPrune(netId);
-        }
-        connectivityChanged = true;
-        break;
-      }
-      case "add_no_connect": {
-        const idExists = [
-          ...draft.instances,
-          ...draft.nets,
-          ...draft.routes,
-          ...draft.junctions,
-          ...draft.noConnects,
-          ...draft.annotations,
-          ...draft.layoutGroups,
-          ...draft.constraints,
-        ].some((candidate) => candidate.id === edit.noConnect.id);
-        if (idExists) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            `Object ID already exists: ${edit.noConnect.id}`,
-          );
-        }
-        draft.noConnects.push(NoConnectSchema.parse(edit.noConnect));
-        changedObjectIds.add(edit.noConnect.id);
-        connectivityChanged = true;
-        break;
-      }
-      case "remove_no_connect": {
-        const index = draft.noConnects.findIndex(
-          (candidate) => candidate.id === edit.noConnectId,
-        );
-        if (index < 0) {
-          return rejectAt(
-            "OBJECT_NOT_FOUND",
-            `NoConnect does not exist: ${edit.noConnectId}`,
-            [],
-            [edit.noConnectId],
-          );
-        }
-        draft.noConnects.splice(index, 1);
-        changedObjectIds.add(edit.noConnectId);
-        connectivityChanged = true;
-        break;
-      }
+      case "add_instance":
+      case "remove_instance":
+      case "add_no_connect":
+      case "remove_no_connect":
       case "set_instance_symbol": {
-        const instance = draft.instances.find(
-          (candidate) => candidate.id === edit.instanceId,
-        );
-        if (!instance) {
-          return rejectAt(
-            "OBJECT_NOT_FOUND",
-            `Instance does not exist: ${edit.instanceId}`,
-            [],
-            [edit.instanceId],
-          );
-        }
-        const lockOwner = lockedLayoutOwner(draft, edit.instanceId);
-        if (lockOwner) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            `Instance ${edit.instanceId} is locked by layout intent ${lockOwner}`,
-          );
-        }
-        const resolver = context.symbolResolver;
-        if (!resolver) {
-          return rejectAt(
-            "EDIT_CONTEXT_REQUIRED",
-            "Symbol edits require a Symbol Resolver",
-          );
-        }
-        const symbolVariantId = edit.symbolVariantId ?? undefined;
-        const resolved = resolver.resolve(edit.symbolId, symbolVariantId);
-        if (!resolved) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            `Symbol or variant does not exist: ${edit.symbolId}${symbolVariantId ? `/${symbolVariantId}` : ""}`,
-          );
-        }
-        const targetPins = new Set(
-          resolved.definition.pins.map((pin) => pin.name),
-        );
-        const pinMap = edit.pinMap ?? {};
-        if (
-          instance.netlist?.binding?.kind === "subcircuit" &&
-          instance.importProvenance?.terminalMapping
-        ) {
-          instance.importProvenance.terminalMapping =
-            instance.importProvenance.terminalMapping.filter((terminal) =>
-              targetPins.has(pinMap[terminal.pinName] ?? terminal.pinName),
-            );
-        }
-        const currentPins = new Set(
-          draft.nets.flatMap((net) =>
-            net.terminals
-              .filter((terminal) => terminal.instanceId === edit.instanceId)
-              .map((terminal) => terminal.pinName),
-          ),
-        );
-        for (const route of draft.routes) {
-          for (const endpoint of [route.from, route.to]) {
-            if (
-              endpoint.kind === "terminal" &&
-              endpoint.instanceId === edit.instanceId
-            ) {
-              currentPins.add(endpoint.pinName);
-            }
-          }
-        }
-        for (const noConnect of draft.noConnects) {
-          if (noConnect.endpoint.instanceId === edit.instanceId) {
-            currentPins.add(noConnect.endpoint.pinName);
-          }
-        }
-        for (const terminal of instance.importProvenance?.terminalMapping ??
-          []) {
-          currentPins.add(terminal.pinName);
-        }
-        for (const sourcePin of Object.keys(pinMap)) {
-          if (!currentPins.has(sourcePin)) {
-            return rejectAt(
-              "EDIT_PRECONDITION",
-              `Pin map source is not connected or routed: ${edit.instanceId}.${sourcePin}`,
-            );
-          }
-        }
-        const mappedPins = new Map<string, string>();
-        for (const sourcePin of currentPins) {
-          const targetPin = pinMap[sourcePin] ?? sourcePin;
-          if (!targetPins.has(targetPin)) {
-            return rejectAt(
-              "EDIT_PRECONDITION",
-              `Target symbol pin does not exist: ${edit.instanceId}.${targetPin}`,
-            );
-          }
-          const previousSource = mappedPins.get(targetPin);
-          if (previousSource && previousSource !== sourcePin) {
-            const ownerNetIds = new Set(
-              draft.nets.flatMap((net) =>
-                net.terminals.some(
-                  (terminal) =>
-                    terminal.instanceId === edit.instanceId &&
-                    (terminal.pinName === previousSource ||
-                      terminal.pinName === sourcePin),
-                )
-                  ? [net.id]
-                  : [],
-              ),
-            );
-            if (ownerNetIds.size > 1) {
-              return rejectAt(
-                "EDIT_PRECONDITION",
-                `Pin map aliases ${previousSource} and ${sourcePin} to ${targetPin} before their Nets are merged`,
-              );
-            }
-          }
-          mappedPins.set(targetPin, sourcePin);
-        }
-        for (const net of draft.nets) {
-          let changed = false;
-          for (const terminal of net.terminals) {
-            if (terminal.instanceId !== edit.instanceId) continue;
-            terminal.pinName = pinMap[terminal.pinName] ?? terminal.pinName;
-            changed = true;
-          }
-          if (changed) changedObjectIds.add(net.id);
-          if (changed) {
-            net.terminals = net.terminals.filter(
-              (terminal, index, terminals) =>
-                terminals.findIndex(
-                  (candidate) =>
-                    candidate.instanceId === terminal.instanceId &&
-                    candidate.pinName === terminal.pinName,
-                ) === index,
-            );
-          }
-        }
-        for (const route of draft.routes) {
-          let changed = false;
-          for (const endpoint of [route.from, route.to]) {
-            if (
-              endpoint.kind === "terminal" &&
-              endpoint.instanceId === edit.instanceId
-            ) {
-              endpoint.pinName = pinMap[endpoint.pinName] ?? endpoint.pinName;
-              changed = true;
-            }
-          }
-          if (changed) changedObjectIds.add(route.id);
-        }
-        for (const noConnect of draft.noConnects) {
-          if (noConnect.endpoint.instanceId !== edit.instanceId) continue;
-          noConnect.endpoint.pinName =
-            pinMap[noConnect.endpoint.pinName] ?? noConnect.endpoint.pinName;
-          changedObjectIds.add(noConnect.id);
-        }
-        if (instance.importProvenance?.terminalMapping) {
-          instance.importProvenance.terminalMapping =
-            instance.importProvenance.terminalMapping.map((terminal) => ({
-              ...terminal,
-              pinName: pinMap[terminal.pinName] ?? terminal.pinName,
-            }));
-        }
-        instance.symbolId = edit.symbolId;
-        if (symbolVariantId === undefined) delete instance.symbolVariantId;
-        else instance.symbolVariantId = symbolVariantId;
-        changedObjectIds.add(instance.id);
-        break;
-      }
-      case "place_instance": {
-        const instance = draft.instances.find(
-          (candidate) => candidate.id === edit.instanceId,
-        );
-        if (!instance) {
-          return rejectAt(
-            "OBJECT_NOT_FOUND",
-            `Instance does not exist: ${edit.instanceId}`,
-            [],
-            [edit.instanceId],
-          );
-        }
-        const lockOwner = lockedLayoutOwner(draft, edit.instanceId);
-        if (lockOwner) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            `Instance ${edit.instanceId} is locked by layout intent ${lockOwner}`,
-          );
-        }
-        if (instance.placement !== null) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            `Instance is already placed: ${edit.instanceId}`,
-          );
-        }
-        instance.placement = structuredClone(edit.placement);
-        changedObjectIds.add(edit.instanceId);
-        break;
-      }
-      case "unplace_instance": {
-        const instance = draft.instances.find(
-          (candidate) => candidate.id === edit.instanceId,
-        );
-        if (!instance) {
-          return rejectAt(
-            "OBJECT_NOT_FOUND",
-            `Instance does not exist: ${edit.instanceId}`,
-            [],
-            [edit.instanceId],
-          );
-        }
-        const lockOwner = lockedLayoutOwner(draft, edit.instanceId);
-        if (lockOwner) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            `Instance ${edit.instanceId} is locked by layout intent ${lockOwner}`,
-          );
-        }
-        if (instance.placement === null) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            `Instance is already unplaced: ${edit.instanceId}`,
-          );
-        }
-        if (
-          draft.routes.some((route) =>
-            [route.from, route.to].some(
-              (endpoint) =>
-                endpoint.kind === "terminal" &&
-                endpoint.instanceId === edit.instanceId,
-            ),
-          )
-        ) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            `Instance has routed terminals; detach routes before unplacing: ${edit.instanceId}`,
-          );
-        }
-        instance.placement = null;
-        changedObjectIds.add(edit.instanceId);
-        break;
-      }
-      case "move_instance": {
-        const instance = draft.instances.find(
-          (candidate) => candidate.id === edit.instanceId,
-        );
-        if (!instance) {
-          return rejectAt(
-            "OBJECT_NOT_FOUND",
-            `Instance does not exist: ${edit.instanceId}`,
-            [],
-            [edit.instanceId],
-          );
-        }
-        const lockOwner = lockedLayoutOwner(draft, edit.instanceId);
-        if (lockOwner) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            `Instance ${edit.instanceId} is locked by layout intent ${lockOwner}`,
-          );
-        }
-        if (instance.placement === null) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            `Instance is not placed: ${edit.instanceId}`,
-          );
-        }
-        const beforeTransform = structuredClone(draft);
-        const oldPlacement = structuredClone(instance.placement);
-        instance.placement.position = structuredClone(edit.position);
-        followAttachedAnnotations(
+        const outcome = applyInstanceLifecycleEdit(edit, {
           draft,
-          edit.instanceId,
-          oldPlacement.position,
-          oldPlacement,
-          instance.placement.position,
-          instance.placement,
+          resolver,
           changedObjectIds,
-          context.symbolResolver,
-        );
-        // Use the snapshot taken immediately before this edit. This is still
-        // progressive for multi-edit transactions, but unlike the old path it
-        // retains the terminal's actual pre-move coordinates.
-        const resolver = context.symbolResolver;
-        if (resolver) {
-          const stretched = applyInstanceRouteFollow(
-            draft,
-            beforeTransform,
-            resolver,
-            edit.instanceId,
-            explicitlyAuthoredRouteIds,
-          );
-          for (const routeId of stretched) changedObjectIds.add(routeId);
-        }
-        changedObjectIds.add(edit.instanceId);
+          deferNetPrune,
+          reject: rejectAt,
+        });
+        if (!outcome.ok) return outcome.rejection;
+        connectivityChanged ||= outcome.connectivityChanged;
         break;
       }
-      case "rotate_instance": {
-        const instance = draft.instances.find(
-          (candidate) => candidate.id === edit.instanceId,
-        );
-        if (!instance) {
-          return rejectAt(
-            "OBJECT_NOT_FOUND",
-            `Instance does not exist: ${edit.instanceId}`,
-            [],
-            [edit.instanceId],
-          );
-        }
-        const lockOwner = lockedLayoutOwner(draft, edit.instanceId);
-        if (lockOwner) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            `Instance ${edit.instanceId} is locked by layout intent ${lockOwner}`,
-          );
-        }
-        if (instance.placement === null) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            `Instance is not placed: ${edit.instanceId}`,
-          );
-        }
-        const beforeTransform = structuredClone(draft);
-        const oldPlacement = structuredClone(instance.placement);
-        instance.placement.rotation = edit.rotation;
-        followAttachedAnnotations(
-          draft,
-          edit.instanceId,
-          oldPlacement.position,
-          oldPlacement,
-          instance.placement.position,
-          instance.placement,
-          changedObjectIds,
-          context.symbolResolver,
-        );
-        const rotateResolver = context.symbolResolver;
-        if (rotateResolver) {
-          for (const routeId of applyInstanceRouteFollow(
-            draft,
-            beforeTransform,
-            rotateResolver,
-            edit.instanceId,
-            explicitlyAuthoredRouteIds,
-          )) {
-            changedObjectIds.add(routeId);
-          }
-        }
-        changedObjectIds.add(edit.instanceId);
-        break;
-      }
+      case "place_instance":
+      case "unplace_instance":
+      case "move_instance":
+      case "rotate_instance":
       case "mirror_instance": {
-        const instance = draft.instances.find(
-          (candidate) => candidate.id === edit.instanceId,
-        );
-        if (!instance) {
-          return rejectAt(
-            "OBJECT_NOT_FOUND",
-            `Instance does not exist: ${edit.instanceId}`,
-            [],
-            [edit.instanceId],
-          );
-        }
-        const lockOwner = lockedLayoutOwner(draft, edit.instanceId);
-        if (lockOwner) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            `Instance ${edit.instanceId} is locked by layout intent ${lockOwner}`,
-          );
-        }
-        if (instance.placement === null) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            `Instance is not placed: ${edit.instanceId}`,
-          );
-        }
-        const beforeTransform = structuredClone(draft);
-        const oldPlacement = structuredClone(instance.placement);
-        instance.placement.mirror = edit.mirror;
-        followAttachedAnnotations(
+        const outcome = applyInstanceTransformEdit(edit, {
           draft,
-          edit.instanceId,
-          oldPlacement.position,
-          oldPlacement,
-          instance.placement.position,
-          instance.placement,
+          resolver,
+          explicitlyAuthoredRouteIds,
           changedObjectIds,
-          context.symbolResolver,
-        );
-        const mirrorResolver = context.symbolResolver;
-        if (mirrorResolver) {
-          for (const routeId of applyInstanceRouteFollow(
-            draft,
-            beforeTransform,
-            mirrorResolver,
-            edit.instanceId,
-            explicitlyAuthoredRouteIds,
-          )) {
-            changedObjectIds.add(routeId);
-          }
-        }
-        changedObjectIds.add(edit.instanceId);
+          reject: rejectAt,
+        });
+        if (!outcome.ok) return outcome.rejection;
         break;
       }
-      case "patch_instance_netlist_parameters": {
-        const instance = draft.instances.find(
-          (candidate) => candidate.id === edit.instanceId,
-        );
-        if (!instance) {
-          return rejectAt(
-            "OBJECT_NOT_FOUND",
-            `Instance does not exist: ${edit.instanceId}`,
-            [],
-            [edit.instanceId],
-          );
-        }
-        const set = edit.set ?? {};
-        const unset = edit.unset ?? [];
-        if (Object.keys(set).length === 0 && unset.length === 0) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            "Netlist parameter patch must set or unset at least one parameter",
-            [],
-            [edit.instanceId],
-          );
-        }
-        const duplicateUnset = new Set(unset);
-        if (duplicateUnset.size !== unset.length) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            "Netlist parameter patch cannot unset the same parameter more than once",
-            [],
-            [edit.instanceId],
-          );
-        }
-        const conflictingKey = Object.keys(set).find((key) =>
-          duplicateUnset.has(key),
-        );
-        if (conflictingKey) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            `Netlist parameter patch cannot set and unset ${conflictingKey}`,
-            [],
-            [edit.instanceId],
-          );
-        }
-        const before: SchematicDocument["instances"][number] =
-          structuredClone(instance);
-        const netlist = instance.netlist;
-        if (!netlist) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            "Netlist parameter patch requires an instance netlist record",
-            [],
-            [edit.instanceId],
-          );
-        }
-        const nextParameters = { ...netlist.parameters };
-        // Delete first so a case-only rename (for example `w` to `W`) is one
-        // valid atomic field change instead of a transient duplicate.
-        for (const key of unset) {
-          delete nextParameters[key];
-        }
-        for (const [key, value] of Object.entries(set)) {
-          nextParameters[key] = value;
-        }
-        const namesByFoldedName = new Map<string, string>();
-        for (const name of Object.keys(nextParameters)) {
-          const foldedName = name.toLowerCase();
-          const prior = namesByFoldedName.get(foldedName);
-          if (prior && prior !== name) {
-            return rejectAt(
-              "EDIT_PRECONDITION",
-              `Netlist parameter ${name} duplicates ${prior} under case folding`,
-              [],
-              [edit.instanceId],
-            );
-          }
-          namesByFoldedName.set(foldedName, name);
-        }
-        const changed =
-          Object.keys(nextParameters).length !==
-            Object.keys(netlist.parameters).length ||
-          Object.entries(nextParameters).some(
-            ([key, value]) => netlist.parameters[key] !== value,
-          );
-        if (!changed) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            "Netlist parameter patch does not change the instance",
-            [],
-            [edit.instanceId],
-          );
-        }
-        netlist.parameters = nextParameters;
-        refreshInstanceValueAnnotation(
-          draft,
-          before,
-          edit.instanceId,
-          changedObjectIds,
-        );
-        changedObjectIds.add(edit.instanceId);
-        break;
-      }
-      case "set_instance_reference": {
-        const instance = draft.instances.find(
-          (candidate) => candidate.id === edit.instanceId,
-        );
-        if (!instance?.netlist) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            "Reference edit requires an instance netlist record",
-            [],
-            [edit.instanceId],
-          );
-        }
-        if (instance.netlist.reference === edit.reference) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            "Reference edit does not change the instance",
-            [],
-            [edit.instanceId],
-          );
-        }
-        instance.netlist.reference = edit.reference;
-        const failure = referencePolicyFailure(draft, instance.id);
-        if (failure) {
-          return rejectAt("EDIT_PRECONDITION", failure, [], [instance.id]);
-        }
-        changedObjectIds.add(edit.instanceId);
-        break;
-      }
-      case "set_instance_schematic_reference": {
-        const instance = draft.instances.find(
-          (candidate) => candidate.id === edit.instanceId,
-        );
-        if (!instance) {
-          return rejectAt(
-            "OBJECT_NOT_FOUND",
-            `Instance does not exist: ${edit.instanceId}`,
-            [],
-            [edit.instanceId],
-          );
-        }
-        if (
-          draft.netlist?.terminals.some((terminal) =>
-            terminal.interfaceInstanceIds.includes(instance.id),
-          )
-        ) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            "A formal Cell Pin is identified by its Cell terminal name, not a schematic reference",
-            [],
-            [instance.id],
-          );
-        }
-        if (instance.schematicReference === edit.reference) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            "Schematic reference edit does not change the instance",
-            [],
-            [edit.instanceId],
-          );
-        }
-        const duplicate = draft.instances.find(
-          (candidate) =>
-            candidate.id !== instance.id &&
-            (
-              candidate.schematicReference ?? candidate.netlist?.reference
-            )?.toLowerCase() === edit.reference.toLowerCase(),
-        );
-        if (duplicate) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            `Schematic reference is already in use: ${edit.reference}`,
-            [],
-            [duplicate.id],
-          );
-        }
-        instance.schematicReference = edit.reference;
-        changedObjectIds.add(instance.id);
-        for (const annotation of draft.annotations) {
-          if (
-            annotation.binding?.kind === "instance-schematic-name" &&
-            annotation.binding.instanceId === instance.id
-          ) {
-            changedObjectIds.add(annotation.id);
-          }
-        }
-        break;
-      }
-      case "set_instance_schematic_name": {
-        const instance = draft.instances.find(
-          (candidate) => candidate.id === edit.instanceId,
-        );
-        if (!instance) {
-          return rejectAt(
-            "OBJECT_NOT_FOUND",
-            `Instance does not exist: ${edit.instanceId}`,
-            [],
-            [edit.instanceId],
-          );
-        }
-        if (
-          JSON.stringify(instance.schematicName ?? null) ===
-          JSON.stringify(edit.content)
-        ) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            "Schematic name edit does not change the instance",
-            [],
-            [edit.instanceId],
-          );
-        }
-        instance.schematicName = structuredClone(edit.content);
-        changedObjectIds.add(edit.instanceId);
-        for (const annotation of draft.annotations) {
-          if (
-            annotation.binding?.kind === "instance-schematic-name" &&
-            annotation.binding.instanceId === instance.id
-          ) {
-            changedObjectIds.add(annotation.id);
-          }
-        }
-        break;
-      }
-      case "set_instance_binding": {
-        const instance = draft.instances.find(
-          (candidate) => candidate.id === edit.instanceId,
-        );
-        if (!instance?.netlist) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            "Binding edit requires an instance netlist record",
-            [],
-            [edit.instanceId],
-          );
-        }
-        const current = instance.netlist.binding ?? null;
-        if (JSON.stringify(current) === JSON.stringify(edit.binding)) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            "Binding edit does not change the instance",
-            [],
-            [edit.instanceId],
-          );
-        }
-        if (edit.binding)
-          instance.netlist.binding = structuredClone(edit.binding);
-        else delete instance.netlist.binding;
-        const failure = referencePolicyFailure(draft, instance.id);
-        if (failure) {
-          return rejectAt("EDIT_PRECONDITION", failure, [], [instance.id]);
-        }
-        changedObjectIds.add(edit.instanceId);
-        connectivityChanged = true;
-        break;
-      }
-      case "set_instance_netlist": {
-        const instance = draft.instances.find(
-          (candidate) => candidate.id === edit.instanceId,
-        );
-        if (!instance) {
-          return rejectAt(
-            "OBJECT_NOT_FOUND",
-            `Instance does not exist: ${edit.instanceId}`,
-            [],
-            [edit.instanceId],
-          );
-        }
-        if (
-          instance.netlist &&
-          JSON.stringify(instance.netlist) === JSON.stringify(edit.netlist)
-        ) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            "Netlist edit does not change the instance",
-            [],
-            [edit.instanceId],
-          );
-        }
-        const before: SchematicDocument["instances"][number] =
-          structuredClone(instance);
-        instance.netlist = structuredClone(edit.netlist);
-        const failure = referencePolicyFailure(draft, instance.id);
-        if (failure) {
-          return rejectAt("EDIT_PRECONDITION", failure, [], [instance.id]);
-        }
-        refreshInstanceValueAnnotation(
-          draft,
-          before,
-          edit.instanceId,
-          changedObjectIds,
-        );
-        changedObjectIds.add(edit.instanceId);
-        connectivityChanged = true;
-        break;
-      }
+      case "patch_instance_netlist_parameters":
+      case "set_instance_reference":
+      case "set_instance_schematic_reference":
+      case "set_instance_schematic_name":
+      case "set_instance_binding":
+      case "set_instance_netlist":
       case "bulk_patch_instance_netlist": {
-        const assignedIds = new Set<string>();
-        for (const assignment of edit.assignments) {
-          if (assignedIds.has(assignment.instanceId)) {
-            return rejectAt(
-              "EDIT_PRECONDITION",
-              `Bulk netlist patch repeats instance ${assignment.instanceId}`,
-              [],
-              [assignment.instanceId],
-            );
-          }
-          assignedIds.add(assignment.instanceId);
-          const instance = draft.instances.find(
-            (candidate) => candidate.id === assignment.instanceId,
-          );
-          if (!instance?.netlist) {
-            return rejectAt(
-              "EDIT_PRECONDITION",
-              `Bulk netlist patch requires a netlist record: ${assignment.instanceId}`,
-              [],
-              [assignment.instanceId],
-            );
-          }
-          const before: SchematicDocument["instances"][number] =
-            structuredClone(instance);
-          let changed = false;
-          let parametersChanged = false;
-          if (
-            assignment.reference !== undefined &&
-            instance.netlist.reference !== assignment.reference
-          ) {
-            instance.netlist.reference = assignment.reference;
-            changed = true;
-          }
-          if (assignment.binding !== undefined) {
-            const current = instance.netlist.binding ?? null;
-            if (
-              JSON.stringify(current) !== JSON.stringify(assignment.binding)
-            ) {
-              if (assignment.binding) {
-                instance.netlist.binding = structuredClone(assignment.binding);
-              } else {
-                delete instance.netlist.binding;
-              }
-              changed = true;
-              connectivityChanged = true;
-            }
-          }
-          const set = assignment.set ?? {};
-          const unset = assignment.unset ?? [];
-          const unsetNames = new Set(unset.map((name) => name.toLowerCase()));
-          if (unsetNames.size !== unset.length) {
-            return rejectAt(
-              "EDIT_PRECONDITION",
-              `Bulk netlist patch repeats an unset parameter on ${instance.id}`,
-              [],
-              [instance.id],
-            );
-          }
-          const conflictingKey = Object.keys(set).find((key) =>
-            unsetNames.has(key.toLowerCase()),
-          );
-          if (conflictingKey) {
-            return rejectAt(
-              "EDIT_PRECONDITION",
-              `Bulk netlist patch cannot set and unset ${conflictingKey}`,
-              [],
-              [instance.id],
-            );
-          }
-          if (Object.keys(set).length > 0 || unset.length > 0) {
-            const nextParameters = { ...instance.netlist.parameters };
-            for (const key of unset) delete nextParameters[key];
-            for (const [key, value] of Object.entries(set)) {
-              nextParameters[key] = value;
-            }
-            const namesByFoldedName = new Map<string, string>();
-            for (const name of Object.keys(nextParameters)) {
-              const folded = name.toLowerCase();
-              const prior = namesByFoldedName.get(folded);
-              if (prior && prior !== name) {
-                return rejectAt(
-                  "EDIT_PRECONDITION",
-                  `Netlist parameter ${name} duplicates ${prior} under case folding`,
-                  [],
-                  [instance.id],
-                );
-              }
-              namesByFoldedName.set(folded, name);
-            }
-            parametersChanged =
-              Object.keys(nextParameters).length !==
-                Object.keys(instance.netlist.parameters).length ||
-              Object.entries(nextParameters).some(
-                ([key, value]) => instance.netlist!.parameters[key] !== value,
-              );
-            if (parametersChanged) {
-              instance.netlist.parameters = nextParameters;
-              changed = true;
-            }
-          }
-          if (!changed) {
-            return rejectAt(
-              "EDIT_PRECONDITION",
-              `Bulk netlist patch does not change ${instance.id}`,
-              [],
-              [instance.id],
-            );
-          }
-          if (parametersChanged) {
-            refreshInstanceValueAnnotation(
-              draft,
-              before,
-              instance.id,
-              changedObjectIds,
-            );
-          }
-          changedObjectIds.add(instance.id);
-        }
-        for (const instanceId of assignedIds) {
-          const failure = referencePolicyFailure(draft, instanceId);
-          if (failure) {
-            return rejectAt("EDIT_PRECONDITION", failure, [], [instanceId]);
-          }
-        }
+        const outcome = applyInstanceNetlistEdit(edit, {
+          draft,
+          changedObjectIds,
+          reject: rejectAt,
+        });
+        if (!outcome.ok) return outcome.rejection;
+        connectivityChanged ||= outcome.connectivityChanged;
         break;
       }
       case "add_cell_terminal":
